Add part of speech parsing
This parses the first part of the mecab output.

At the moment, this only handles Nouns, Verbs, and Particles. Further tests
using the entire kokoro sample text should be introduced to ensure that we can parse every mecab output
for the book, but that can wait until we're parsing all the parts mecab outputs.